+++ RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 ### RELEASE 1.1.12
 
+FIXES:
+
+1. GNSSStreamer default input handler updated - Fixes [#111](https://github.com/semuconsulting/pygnssutils/issues/111)
+
 CHANGES:
 
 1. Inherit SocketWrapper from pynmeagps.